pyproject.toml
[project]
name = "agentcore-cx-demo"
version = "0.1.0"
description = "Multi-agent healthcare CX demo — Strands + A2A + AgentCore"
requires-python = ">=3.12"
dependencies = [
    "strands-agents>=1.0.0",
    "strands-agents-tools[a2a]>=0.1.0",
    "bedrock-agentcore>=0.1.0",
    "boto3>=1.35.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.23.0"]
